+++ src/compiler/mod.rs
@@ -1,1 +1,2 @@
 pub(crate) mod codegen;
+pub(crate) mod helpers;
